requirements.txt
picamera
qrcode
gpio4
ConfigParser
pygame
pycups
Image
Pillow-PIL
pil-compat
RPi.GPIO
filelock
textile
coloredlogs